Refactor glove firmware for ADS1115 and BNO085 diagnostics
- Removed obsolete PowerShell script for ESP32-S3 reset and capture.
- Updated main firmware to focus on ADS1115 diagnostics, including I2C scanning and configuration register checks.
- Added new test files for BNO085 diagnostics, including I2C scan and sensor data streaming.
- Created README documentation for BNO085 test suite with wiring instructions and troubleshooting tips.
- Implemented detailed logging for sensor initialization and data retrieval.

+++ glove_firmware/test/test_ads1115/test_ads1115_diagnostic.ino
@@ -81,7 +81,7 @@ void setup() {
     Wire.beginTransmission(ADS1115_ADDR_GND);
     Wire.write(ADS1115_REG_CFG);
     Wire.endTransmission();
-    Wire.requestFrom((int)ADS1115_ADDR_GND, (int)2);
+    Wire.requestFrom(ADS1115_ADDR_GND, (uint8_t)2);
     if (Wire.available() >= 2) {
         uint16_t cfg = (Wire.read() << 8) | Wire.read();
         Serial.printf("[ADS1115] 0x48 config=0x%04X", cfg);
@@ -95,7 +95,7 @@ void setup() {
     Wire.beginTransmission(ADS1115_ADDR_VDD);
     Wire.write(ADS1115_REG_CFG);
     Wire.endTransmission();
-    Wire.requestFrom((int)ADS1115_ADDR_VDD, (int)2);
+    Wire.requestFrom(ADS1115_ADDR_VDD, (uint8_t)2);
     if (Wire.available() >= 2) {
         uint16_t cfg = (Wire.read() << 8) | Wire.read();
         Serial.printf("[ADS1115] 0x49 config=0x%04X", cfg);
